Pyhton3: use 2to3 and fix setup to work with 2 and 3
Remove the import and use of pyhthon-creole. The Readme is still creole
formated and will be harder to read. But 2to3 can't handle incompatibilities
in setup code, and importing creole with custom exception handling during
setup.py just to write the Readme in creole is overkill anyway.

Also remove exception handling for authors so we don't have to code around
it. There is not much that can go wrong here.